refactor: move all attribute tests to a dedicated module
